+++ server/src/routes/bins/bin.routes.ts
@@ -1,7 +1,7 @@
 import express from 'express'
 import { getBin, getBinCodes, getBins } from './bin.controller'
 import roleAllow from 'middlewares/roleAllow.middleware'
-import { UserRole } from '../../constants/roles'
+import { UserRole } from '../../constants/UserRole'
 
 const router = express.Router()
