+++ server/src/routes/tasks/task.controller.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from 'express'
 import * as taskService from '../tasks/task.service'
 import * as binService from 'routes/bins/bin.service'
 import AppError from 'utils/appError'
-import { UserRole } from '../../constants/roles'
+import { UserRole } from '../../constants/UserRole'
 
 export const acceptTask = async (
   req: Request,
